add close project menu on mobile -first steps
